Feat/swagger api documentation

+++ app/Http/Controllers/EmailsController.php
@@ -5,26 +5,35 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreCreatedEmailRequest;
 use App\Models\Email;
 use Illuminate\Http\Request;
+use Illuminate\Http\JsonResponse;
 
 class EmailsController extends Controller
 {
-	public function create(StoreCreatedEmailRequest $request)
+	public function create(StoreCreatedEmailRequest $request): JsonResponse
 	{
+		if (!jwtUser())
+		{
+			return response()->json(['message' => 'token not present'], 401);
+		}
 		$email = Email::create([
 			'email'   => $request->email,
 			'user_id' => jwtUser()->id,
 		]);
 		jwtUser()->sendEmailVerification($email->id, $email->email);
+		return response()->json('Email created successfully!', 201);
 	}
 
 	public function verify($email_id, Request $request)
 	{
 		if (!$request->hasValidSignature())
 		{
-			return response()->json(['msg' => 'Invalid/Expired url provided.'], 401);
+			return response()->json(['msg' => 'Invalid/Expired url provided.'], 410);
 		}
 		$email = Email::findOrFail($email_id);
-
+		if (!$email)
+		{
+			return response()->json(['msg' => 'Email not found'], 404);
+		}
 		if (!$email->is_email_verified)
 		{
 			$email->is_email_verified = 1;
@@ -54,13 +63,13 @@ class EmailsController extends Controller
 
 	public function destroy(Email $email)
 	{
+		if ($email->primary === 1)
+		{
+			return response()->json('Deleting primary email is not allowed', 405);
+		}
 		if ($email->delete())
 		{
-			return response()->json(['email deleted successfully'], 200);
-		}
-		else
-		{
-			return response()->json(['Email deleting film'], 404);
+			return response()->json(['msg' => 'email deleted successfully'], 202);
 		}
 	}
 }
